+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 
 ### Added
 
+- Add a user-visible route to delete a subsection
+
 ### Changed
 
 - Redirect logged-in users from the homepage to the index page